requirements.txt
aiogram>=3.22.0
motor>=3.3.2
emergentintegrations
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiohttp>=3.12.0
apscheduler>=3.10.0
